Appwrite 1.2.0 support
