--- docs/examples/locale/get-languages.md
@@ -1,18 +1,0 @@
-import { Client, Locale } from "appwrite";
-
-const client = new Client();
-
-const locale = new Locale(client);
-
-client
-    .setEndpoint('https://[HOSTNAME_OR_IP]/v1') // Your API Endpoint
-    .setProject('5df5acd0d48c2') // Your project ID
-;
-
-const promise = locale.getLanguages();
-
-promise.then(function (response) {
-    console.log(response); // Success
-}, function (error) {
-    console.log(error); // Failure
-});
