+++ src/client.ts
@@ -102,7 +102,7 @@ class Client {
         'x-sdk-name': 'Web',
         'x-sdk-platform': 'client',
         'x-sdk-language': 'web',
-        'x-sdk-version': '10.1.0',
+        'x-sdk-version': '10.2.0',
         'X-Appwrite-Response-Format': '1.0.0',
     };
