+++ src/services/account.ts
@@ -84,7 +84,7 @@ export class Account extends Service {
         }
 
         /**
-         * Update Account Email
+         * Update Email
          *
          * Update currently logged in user account email address. After changing user
          * address, the user confirmation status will get reset. A new confirmation
@@ -127,7 +127,7 @@ export class Account extends Service {
         }
 
         /**
-         * Create Account JWT
+         * Create JWT
          *
          * Use this endpoint to create a JSON Web Token. You can use the resulting JWT
          * to authenticate on behalf of the current user when working with the
@@ -149,7 +149,7 @@ export class Account extends Service {
         }
 
         /**
-         * List Account Logs
+         * List Logs
          *
          * Get currently logged in user list of latest security activity logs. Each
          * log returns user IP address, location and date and time of log.
@@ -173,7 +173,7 @@ export class Account extends Service {
         }
 
         /**
-         * Update Account Name
+         * Update Name
          *
          * Update currently logged in user account name.
          *
@@ -200,7 +200,7 @@ export class Account extends Service {
         }
 
         /**
-         * Update Account Password
+         * Update Password
          *
          * Update currently logged in user password. For validation, user is required
          * to pass in the new password, and the old password. For users created with
@@ -234,7 +234,7 @@ export class Account extends Service {
         }
 
         /**
-         * Update Account Phone
+         * Update Phone
          *
          * Update the currently logged in user's phone number. After updating the
          * phone number, the phone verification status will be reset. A confirmation
@@ -292,7 +292,7 @@ export class Account extends Service {
         }
 
         /**
-         * Update Account Preferences
+         * Update Preferences
          *
          * Update currently logged in user account preferences. The object you pass is
          * stored as is, and replaces any previous value. The maximum allowed prefs
@@ -426,7 +426,7 @@ export class Account extends Service {
         }
 
         /**
-         * List Account Sessions
+         * List Sessions
          *
          * Get currently logged in user list of active sessions across different
          * devices.
@@ -445,7 +445,7 @@ export class Account extends Service {
         }
 
         /**
-         * Delete All Account Sessions
+         * Delete Sessions
          *
          * Delete all sessions from the user account and remove any sessions cookies
          * from the end client.
@@ -487,10 +487,13 @@ export class Account extends Service {
         }
 
         /**
-         * Create Account Session with Email
+         * Create Email Session
          *
          * Allow the user to login into their account by providing a valid email and
          * password combination. This route will create a new session for the user.
+         * 
+         * A user is limited to 10 active sessions at a time by default. [Learn more
+         * about session limits](/docs/authentication#limits).
          *
          * @param {string} email
          * @param {string} password
@@ -537,6 +540,9 @@ export class Account extends Service {
          * address is valid for 1 hour. If you are on a mobile device you can leave
          * the URL parameter empty, so that the login completion will be handled by
          * your Appwrite instance by default.
+         * 
+         * A user is limited to 10 active sessions at a time by default. [Learn more
+         * about session limits](/docs/authentication#limits).
          *
          * @param {string} userId
          * @param {string} email
@@ -621,7 +627,7 @@ export class Account extends Service {
         }
 
         /**
-         * Create Account Session with OAuth2
+         * Create OAuth2 Session
          *
          * Allow the user to login to their account using the OAuth2 provider of their
          * choice. Each OAuth2 provider should be enabled from the Appwrite console
@@ -633,7 +639,10 @@ export class Account extends Service {
          * attempt to look for a user with the same email address as the email
          * received from the OAuth2 provider and attach the new session to the
          * existing user. If no matching user is found - the server will create a new
-         * user..
+         * user.
+         * 
+         * A user is limited to 10 active sessions at a time by default. [Learn more
+         * about session limits](/docs/authentication#limits).
          * 
          *
          * @param {string} provider
@@ -686,6 +695,9 @@ export class Account extends Service {
          * /account/sessions/phone](/docs/client/account#accountUpdatePhoneSession)
          * endpoint to complete the login process. The secret sent to the user's phone
          * is valid for 15 minutes.
+         * 
+         * A user is limited to 10 active sessions at a time by default. [Learn more
+         * about session limits](/docs/authentication#limits).
          *
          * @param {string} userId
          * @param {string} phone
@@ -759,7 +771,7 @@ export class Account extends Service {
         }
 
         /**
-         * Get Session By ID
+         * Get Session
          *
          * Use this endpoint to get a logged in user's session using a Session ID.
          * Inputting 'current' will return the current session being used.
@@ -783,7 +795,7 @@ export class Account extends Service {
         }
 
         /**
-         * Update Session (Refresh Tokens)
+         * Update OAuth Session (Refresh Tokens)
          *
          * Access tokens have limited lifespan and expire to mitigate security risks.
          * If session was created using an OAuth provider, this route can be used to
@@ -808,7 +820,7 @@ export class Account extends Service {
         }
 
         /**
-         * Delete Account Session
+         * Delete Session
          *
          * Use this endpoint to log out the currently logged in user from all their
          * account sessions across all of their different devices. When using the
@@ -834,7 +846,7 @@ export class Account extends Service {
         }
 
         /**
-         * Update Account Status
+         * Update Status
          *
          * Block the currently logged in user account. Behind the scene, the user
          * record is not deleted but permanently blocked from any access. To
