+++ src/services/databases.ts
@@ -14,9 +14,7 @@ export class Databases extends Service {
          * List Documents
          *
          * Get a list of all the user's documents in a given collection. You can use
-         * the query params to filter your results. On admin mode, this endpoint will
-         * return a list of all of documents belonging to the provided collectionId.
-         * [Learn more about different API modes](/docs/admin).
+         * the query params to filter your results.
          *
          * @param {string} databaseId
          * @param {string} collectionId
