+++ src/services/functions.ts
@@ -11,44 +11,10 @@ export class Functions extends Service {
      }
 
         /**
-         * Retry Build
-         *
-         *
-         * @param {string} functionId
-         * @param {string} deploymentId
-         * @param {string} buildId
-         * @throws {AppwriteException}
-         * @returns {Promise}
-         */
-        async retryBuild(functionId: string, deploymentId: string, buildId: string): Promise<{}> {
-            if (typeof functionId === 'undefined') {
-                throw new AppwriteException('Missing required parameter: "functionId"');
-            }
-
-            if (typeof deploymentId === 'undefined') {
-                throw new AppwriteException('Missing required parameter: "deploymentId"');
-            }
-
-            if (typeof buildId === 'undefined') {
-                throw new AppwriteException('Missing required parameter: "buildId"');
-            }
-
-            let path = '/functions/{functionId}/deployments/{deploymentId}/builds/{buildId}'.replace('{functionId}', functionId).replace('{deploymentId}', deploymentId).replace('{buildId}', buildId);
-            let payload: Payload = {};
-
-            const uri = new URL(this.client.config.endpoint + path);
-            return await this.client.call('post', uri, {
-                'content-type': 'application/json',
-            }, payload);
-        }
-
-        /**
          * List Executions
          *
          * Get a list of all the current user function execution logs. You can use the
-         * query params to filter your results. On admin mode, this endpoint will
-         * return a list of all of the project's executions. [Learn more about
-         * different API modes](/docs/admin).
+         * query params to filter your results.
          *
          * @param {string} functionId
          * @param {string[]} queries
